Add scale module with all functions to manage the scales
